+++ benches/micro.rs
@@ -54,7 +54,7 @@ fn pll_bench() {
 
 fn iir_int_bench() {
     let dut = iir::Biquad::default();
-    let mut xy = [0; 5];
+    let mut xy = [0; 4];
     println!(
         "int_iir::IIR::update(s, x): {}",
         bench_env(0x2832, |x| dut.update(&mut xy, *x))
@@ -63,7 +63,7 @@ fn iir_int_bench() {
 
 fn iir_f32_bench() {
     let dut = iir::Biquad::<f32>::default();
-    let mut xy = [0.0; 5];
+    let mut xy = [0.0; 4];
     println!(
         "int::IIR::<f32>::update(s, x): {}",
         bench_env(0.32241, |x| dut.update(&mut xy, *x))
@@ -72,7 +72,7 @@ fn iir_f32_bench() {
 
 fn iir_f64_bench() {
     let dut = iir::Biquad::<f64>::default();
-    let mut xy = [0.0; 5];
+    let mut xy = [0.0; 4];
     println!(
         "int::IIR::<f64>::update(s, x): {}",
         bench_env(0.32241, |x| dut.update(&mut xy, *x))